Stop after one function iteration when return value is the same function with the same arguments (e.g. when function definitions is f(x)=f(x)); Do not show single input decimal number as only combined fraction if fraction format is set to dual; Apply fraction format also for integers if set to percent, permille, or permyriad; Add (hidden) support for percent, permille, and permyriad fraction formats in qalc; Fix "where" expression when variable name is placed after equals greater or equals less; Fix for "do not show single input decimal number as fraction by default if small or large enough to be displayed using scientific notation"; Fix typo in PrintOptions::use_unicode_signs documentation; Fix compiler warning; Update ChangeLog

+++ libqalculate/MathStructure-integrate.cc
@@ -7490,7 +7490,7 @@ bool MathStructure::integrate(const MathStructure &lower_limit, const MathStruct
 		}
 		eo2.approximation = eo.approximation;
 		if(b) {
-			if(definite_integral && mstruct.containsFunctionId(FUNCTION_ID_INTEGRATE, true) <= 0 && ((den_check < 0 && count_ln(mstruct) <= -den_check) || test_definite_ln(mstruct, m_interval, x_var, eo))) {
+			if(definite_integral && mstruct.containsFunctionId(FUNCTION_ID_INTEGRATE, true) <= 0 && ((den_check < 0 && count_ln(mstruct) <= (size_t) -den_check) || test_definite_ln(mstruct, m_interval, x_var, eo))) {
 				CALCULATOR->endTemporaryStopMessages(true);
 				MathStructure mstruct_lower(mstruct);
 				if(m1.isInfinite() || m2.isInfinite()) {
